new models: plant, plant types, legal types

+++ models/contribution.go
@@ -3,7 +3,7 @@ package models
 import "github.com/astaxie/beego/orm"
 
 type Contribution struct {
-	Id             uint32 `json:"id"`
+	Id             int64 `json:"id"`
 	Code           uint16 `json:"code"`
 	Name           string `json:"name"`
 	InspectionCriteria []*InspectionCriterion `json:"inspectionCriteria" orm:"reverse(many)"`
